double operations changed to float in the math lib for more efficiency

+++ src/the_shadow_realm/math.cpp
@@ -12,13 +12,13 @@ inline double rad2deg(double radians)
 
 // VEC2
 
-Vec2::Vec2(double x, double y)
+Vec2::Vec2(float x, float y)
 {
 	this->x = x;
 	this->y = y;
 }
 
-double Vec2::lenght()
+float Vec2::lenght()
 {
 	return sqrt(x * x + y * y);
 }
@@ -28,17 +28,17 @@ Vec2 Vec2::normalize()
 	return Vec2(x/lenght(), y/lenght());
 }
 
-double Vec2::distanceTo(Vec2& other)
+float Vec2::distanceTo(Vec2& other)
 {
 	return sqrt(pow(x - other.x, 2) + pow(y - other.y, 2));
 }
 
-Vec2 Vec2::rotate(double degrees)
+Vec2 Vec2::rotate(float degrees)
 {
-	double rad = deg2rad(degrees);
-	double aCos = cos(rad);
-	double aSin = sin(rad);
-	return Vec2((x * aCos - y * aSin), (x * aSin * y * aCos));
+	float rad = deg2rad(degrees);
+	float aCos = cos(rad);
+	float aSin = sin(rad);
+	return Vec2((x * (float)aCos - y * (float)aSin), (x * (float)aSin * y * (float)aCos));
 }
 
 Vec2 operator+(const Vec2& a, const Vec2& b)
@@ -51,31 +51,31 @@ Vec2 operator-(const Vec2& a, const Vec2& b)
 	return Vec2(a.x - b.x, a.y - b.y);
 }
 
-Vec2 operator*(const Vec2& a, const double real)
+Vec2 operator*(const Vec2& a, const float real)
 {
 	return Vec2(a.x * real, a.y * real);
 }
 
-Vec2 operator/(const Vec2& a, const double real)
+Vec2 operator/(const Vec2& a, const float real)
 {
 	return Vec2(a.x / real, a.y / real);
 }
 
-double operator*(const Vec2& a, const Vec2& b)
+float operator*(const Vec2& a, const Vec2& b)
 {
 	return (a.x * b.x) + (a.y * b.y);
 }
 
 // VEC3
 
-Vec3::Vec3(double x, double y, double z)
+Vec3::Vec3(float x, float y, float z)
 {
 	this->x = x;
 	this->y = y;
 	this->z = z;
 }
 
-double Vec3::lenght()
+float Vec3::lenght()
 {
 	return sqrt(x * x + y * y + z * z);
 }
@@ -85,16 +85,16 @@ Vec3 Vec3::normalize()
 	return Vec3(x/lenght(), y/lenght(), z/lenght());
 }
 
-double Vec3::distanceTo(Vec3& other)
+float Vec3::distanceTo(Vec3& other)
 {
 	return sqrt(pow(x - other.x, 2) + pow(y - other.y, 2) + pow(z - other.z, 2));
 }
 
 Vec3 Vec3::cross(Vec3& other)
 {
-	double x_ = y * other.z - z * other.y;
-	double y_ = z * other.x - x * other.z;
-	double z_ = x * other.y - y * other.x;
+	float x_ = y * other.z - z * other.y;
+	float y_ = z * other.x - x * other.z;
+	float z_ = x * other.y - y * other.x;
 	return Vec3(x_, y_, z_);
 }
 
@@ -108,29 +108,29 @@ Vec3 operator-(const Vec3& a, const Vec3& b)
 	return Vec3(a.x - b.x, a.y - b.y, a.z - b.z);
 }
 
-Vec3 operator*(const Vec3& a, const double real)
+Vec3 operator*(const Vec3& a, const float real)
 {
 	return Vec3(a.x * real, a.y * real, a.z * real);
 }
 
-Vec3 operator/(const Vec3& a, const double real)
+Vec3 operator/(const Vec3& a, const float real)
 {
 	return Vec3(a.x / real, a.y / real, a.z / real);
 }
 
 // MAT4
 
-double operator*(const Vec3& a, const Vec3& b)
+float operator*(const Vec3& a, const Vec3& b)
 {
 	return (a.x * b.x) + (a.y * b.y) + (a.z * b.z);
 }
 
-double const* Mat4::operator[](int const in) const
+float const* Mat4::operator[](int const in) const
 {
 	return &m[0][in];
 }
 
-double* Mat4::operator[](int const in)
+float* Mat4::operator[](int const in)
 {
 	return &m[0][in];
 }
@@ -158,7 +158,7 @@ void makeIdentity(Mat4& matrix)
 
 // QUATERNION
 
-Quaternion::Quaternion(double x, double y, double z, double w)
+Quaternion::Quaternion(float x, float y, float z, float w)
 {
 	this->x = x;
 	this->y = y;
@@ -166,7 +166,7 @@ Quaternion::Quaternion(double x, double y, double z, double w)
 	this->w = w;
 }
 
-double Quaternion::lenght()
+float Quaternion::lenght()
 {
 	return sqrt(x * x + y * y + z * z + w * w);
 }
